added render function for server
