+++ src/App.jsx
@@ -1,19 +1,29 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Switch, Route, NavLink } from 'react-router-dom';
 
 // base SASS file
 import './styles/App.scss';
 
+const Header = () => {
+  useEffect(() => {
+    console.log('salam');
+  }, []);
+
+  return (
+    <header className="site-header">
+      <NavLink to="/" exact>
+        Home
+      </NavLink>
+      <NavLink to="/about">About</NavLink>
+      <NavLink to="/projects"> Projects </NavLink>
+    </header>
+  );
+};
+
 const App = ({ routes, initialData }) => {
   return routes ? (
     <div>
-      <header className="site-header">
-        <NavLink to="/" exact>
-          Home
-        </NavLink>
-        <NavLink to="/about">About</NavLink>
-        <NavLink to="/projects"> Projects </NavLink>
-      </header>
+      <Header />
       <main className="main-site">
         <Switch>
           {routes.map((routeItem, index) => (
